chore(kit): refactor `min` / `max` logic for new `InputMonth`
