+++ projects/demo/src/modules/components/input-month/examples/5/index.html
@@ -3,8 +3,6 @@
         placeholder="Use mobile device"
         tuiInputMonth
         type="month"
-        [max]="max"
-        [min]="min"
         [(ngModel)]="value"
     />
