+++ projects/demo/src/modules/components/input-month/examples/5/index.ts
@@ -2,7 +2,7 @@ import {Component} from '@angular/core';
 import {FormsModule} from '@angular/forms';
 import {changeDetection} from '@demo/emulate/change-detection';
 import {encapsulation} from '@demo/emulate/encapsulation';
-import {TuiMonth} from '@taiga-ui/cdk';
+import type {TuiMonth} from '@taiga-ui/cdk';
 import {TuiTextfield} from '@taiga-ui/core';
 import {TuiInputMonth} from '@taiga-ui/kit';
 
@@ -14,7 +14,5 @@ import {TuiInputMonth} from '@taiga-ui/kit';
     changeDetection,
 })
 export default class Example {
-    protected readonly min = TuiMonth.currentLocal().append({month: -12});
-    protected readonly max = TuiMonth.currentLocal().append({month: 12});
     protected value: TuiMonth | null = null;
 }
